fix: don't infer creation code and inject it
Using a INIT_CODE_PAIR_HASH can yield unexpected results because it's
not part of uniswap v2 contract. Hence not all AMMs comply to it
generating side effects

We inject the creationCode alongside the factory on contract creation or
mutation. Bear in mind we are not validating any of the two (factory
address / creationCode) so manual input can still pose a risk